GafferSceneUI bindings : Reorganise for simplicity and compilation speed
