+++ src/GafferSceneUIModule/VisualiserBinding.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2016, John Haddon. All rights reserved.
+//  Copyright (c) 2015, John Haddon. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -34,24 +34,14 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
-#include "GafferBindings/NodeBinding.h"
+#ifndef GAFFERSCENEUIMODULE_VISUALISERBINDING_H
+#define GAFFERSCENEUIMODULE_VISUALISERBINDING_H
 
-#include "GafferUI/Gadget.h"
-
-#include "GafferSceneUI/SceneView.h"
-#include "GafferSceneUI/ScaleTool.h"
-#include "GafferSceneUIBindings/ScaleToolBinding.h"
-
-using namespace boost::python;
-using namespace GafferSceneUI;
-
-void GafferSceneUIBindings::bindScaleTool()
+namespace GafferSceneUIModule
 {
 
-	scope s = GafferBindings::NodeClass<ScaleTool>( NULL, no_init )
-		.def( init<SceneView *>() )
-		.def( "scale", &ScaleTool::scale )
-	;
+void bindVisualisers();
 
-}
+} // namespace GafferSceneUIModule
 
+#endif // GAFFERSCENEUIMODULE_VISUALISERBINDING_H
